// src/api/models/ticketProductResDTOType.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type TicketProductResDTOType = typeof TicketProductResDTOType[keyof typeof TicketProductResDTOType];


export const TicketProductResDTOType = {
  EXPERIENCE: 'EXPERIENCE',
  PORTFOLIO_CORRECTION: 'PORTFOLIO_CORRECTION',
} as const;