// src/api/models/paymentResDTOStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type PaymentResDTOStatus = typeof PaymentResDTOStatus[keyof typeof PaymentResDTOStatus];


export const PaymentResDTOStatus = {
  REQUESTED: 'REQUESTED',
  WAITING: 'WAITING',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED',
  PARTIAL_REFUNDED: 'PARTIAL_REFUNDED',
} as const;